pyproject.toml
[project]
name = "guardial-garden"
version = "1.0.0"
description = "Guardil CRUD Repository"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi==0.127.0",
    "uvicorn==0.40.0",
    "azure-cosmos",
    "azure-identity",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "prompty>=0.1.0",
    "jinja2>=3.1.0",
    "PyJWT==2.10.1",
    "azure-ai-contentsafety==1.0.0",
    "azure-ai-textanalytics==5.2.0",
    "pymongo==4.16.0",
    "python-multipart==0.0.22"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  \.git
  | \.venv
  | venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",
    "F",
    "W",
    "I",
    "N",
    "UP",
    "YTT",
    "ASYNC",
    "S",
    "BLE",
    "B",
    "A",
    "C4",
    "T10",
    "ISC",
    "ICN",
    "G",
    "PIE",
    "PYI",
    "RSE",
    "RET",
    "SIM",
    "TID",
    "ARG",
    "PTH",
    "PD",
    "PGH",
    "PL",
    "TRY",
    "RUF",
]
ignore = [
    "S101",
    "TRY003",
    "PLR0913",
    "UP007",
]

[tool.ruff.per-file-ignores]
"tests/**/*.py" = ["S101", "PLR2004", "ARG001"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
addopts = "-v --strict-markers --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
